+++ test/apns_config_test.exs
@@ -1,0 +1,9 @@
+defmodule Pigeon.APNS.ConfigTest do
+  use ExUnit.Case
+
+  alias Pigeon.APNS.Config
+
+  test "valid? returns true if proper ssl config keys present" do
+    assert Config.valid?(Config.config(:default))
+  end
+end